+++ source/be_pcre.c
@@ -69,10 +69,10 @@ object compile(object pattern, object eflags) {
 	}
 	
 	
-	if ((unsigned) re > (unsigned)MAXINT)
-		ret = NewDouble((double)(unsigned long)re);
+	if ((uintptr_t) re > (uintptr_t)MAXINT)
+		ret = NewDouble((double)(uintptr_t)re);
 	else
-		ret = (unsigned long)re;
+		ret = (uintptr_t)re;
 	
 	return ret;
 }
@@ -134,7 +134,7 @@ object compile_pcre(object x, object flags) {
 			rcp->re = (struct real_pcre *) compiled_regex;
 		}
 		else {
-			rcp->re = (struct real_pcre *) (unsigned long) DBL_PTR( compiled_regex )->dbl;
+			rcp->re = (struct real_pcre *) (uintptr_t) DBL_PTR( compiled_regex )->dbl;
 		}
 		x = MAKE_SEQ( regex );
 	}
